Move converters to its namespace
